refactor(sync): extract server dialog helpers
